[MOD/#386] Refactor calendar util and model files
- Move `DateExtensions.kt` from model to util package.
- Remove unused imports in `MonthData.kt`.
- Replace `Exception()` with `error()` for unexpected month date in `MonthData.kt`.
- Change internal preview function to private in `DaysOfWeekTitle.kt`.
